Marker API (GPU)

Note: This is a maturing feature. Only NVIDIA GPUs are supported.

Example

(See https://github.com/JuliaPerf/LIKWID.jl/tree/main/examples/perfctr_gpu.)

# perfctr_gpu.jl
using LIKWID
using LinearAlgebra
using CUDA

@assert CUDA.functional()

GPUMarker.init()

# Note: CUDA defaults to Float32
Agpu = CUDA.rand(128, 64)
Bgpu = CUDA.rand(64, 128)
Cgpu = CUDA.zeros(128, 128)

GPUMarker.startregion("matmul")
for _ in 1:100
    mul!(Cgpu, Agpu, Bgpu)
end
GPUMarker.stopregion("matmul")

GPUMarker.close()

Running this file with the command likwid-perfctr -G 0 -W FLOPS_SP -m julia perfctr_gpu.jl one should obtain something like the following:

--------------------------------------------------------------------------------
CPU name:	Intel(R) Xeon(R) Gold 6246 CPU @ 3.30GHz
CPU type:	Intel Cascadelake SP processor
CPU clock:	3.30 GHz
--------------------------------------------------------------------------------
--------------------------------------------------------------------------------
Region matmul, Group 1: FLOPS_SP
+-------------------+----------+
|    Region Info    |   GPU 0  |
+-------------------+----------+
| RDTSC Runtime [s] | 3.426146 |
|     call count    |        1 |
+-------------------+----------+
+----------------------------------------------------+---------+----------+
|                        Event                       | Counter |   GPU 0  |
+----------------------------------------------------+---------+----------+
| SMSP_SASS_THREAD_INST_EXECUTED_OP_FADD_PRED_ON_SUM |   GPU0  |  2457600 |
| SMSP_SASS_THREAD_INST_EXECUTED_OP_FMUL_PRED_ON_SUM |   GPU1  |  3276800 |
| SMSP_SASS_THREAD_INST_EXECUTED_OP_FFMA_PRED_ON_SUM |   GPU2  | 52436990 |
+----------------------------------------------------+---------+----------+
+---------------------+---------+
|        Metric       |  GPU 0  |
+---------------------+---------+
| Runtime (RDTSC) [s] |  3.4261 |
|     SP [MFLOP/s]    | 32.2836 |
+---------------------+---------+

Convenience macro

We provide (and export) the macro @gpuregion which can be used to write regions like

GPUMarker.startregion("matmul")
for _ in 1:100
    mul!(Cgpu, Agpu, Bgpu)
end
GPUMarker.stopregion("matmul")

simply as

@gpuregion "matmul" for _ in 1:100
    mul!(Cgpu, Agpu, Bgpu)
end

likwid-perfctr in a nutshell

Most importantly, as for CPUs, you need to use the -m option to activate the marker API.

To list the available GPU performance groups, run likwid-perfctr -a and look for the lower "NvMon" table:

[...]

NvMon group name	Description
--------------------------------------------------------------------------------
    DATA	Load to store ratio
FLOPS_SP	Single-precision floating point
FLOPS_HP	Half-precision floating point
FLOPS_DP	Double-precision floating point

These groups can be passed to the command line option -W. Note that you can also query the available GPU performance groups programmatically using LIKWID.NvMon.get_groups(gpuid::Integer).

Another important option is -G <list>, where <list> is a list of GPUs to monitor. Note that GPU ids start with zero (not one).

Combinding the points above, the full command could look like this: likwid-perfctr -G 0 -W FLOPS_SP -m julia.

For more information, check out the official documentation.

API

LIKWID.GPUMarker.close

LIKWID.GPUMarker.getregion

LIKWID.GPUMarker.gpuregion

LIKWID.GPUMarker.init

LIKWID.GPUMarker.isactive

LIKWID.GPUMarker.nextgroup

LIKWID.GPUMarker.registerregion

LIKWID.GPUMarker.resetregion

LIKWID.GPUMarker.startregion

LIKWID.GPUMarker.stopregion

Functions

LIKWID.GPUMarker.close — Method

Close the connection to the LIKWID GPU Marker API and write out measurement data to file. This file will be evaluated by likwid-perfctr.

LIKWID.GPUMarker.getregion — Method

getregion(regiontag::AbstractString) -> nevents, events, time, count

Get the intermediate results of the region identified by regiontag. On success, it returns * nevents: the number of events in the current group, * events: a list with all the aggregated event results, * time: the measurement time for the region and * count: the number of calls.

LIKWID.GPUMarker.gpuregion — Method

gpuregion(f, regiontag::AbstractString)

Adds a LIKWID GPU marker region around the execution of the given function f using GPUMarker.startregion, GPUMarker.stopregion under the hood. Note that LIKWID.GPUMarker.init() and LIKWID.GPUMarker.close() must be called before and after, respectively.

Examples

julia> using LIKWID, CUDA

julia> GPUMarker.init()

julia> gpuregion("sleeping...") do
           sleep(1)
       end
true

julia> gpuregion(()->CUDA.rand(100), "create rand vec")
true

julia> GPUMarker.close()

LIKWID.GPUMarker.init — Method

Initialize the NvMon Marker API of the LIKWID library. Must be called previous to all other functions.

LIKWID.GPUMarker.isactive — Method

Checks whether the NVIDIA GPU Marker API is active, i.e. julia has been started under likwid-perfctr -G ... -W ... -m.

LIKWID.GPUMarker.nextgroup — Method

Switch to the next event set in a round-robin fashion. If you have set only one event set on the command line, this function performs no operation.

LIKWID.GPUMarker.registerregion — Method

Register a region with name regiontag to the GPU Marker API. On success, true is returned.

This is an optional function to reduce the overhead of region registration at Marker.startregion. If you don't call registerregion, the registration is done at startregion.

LIKWID.GPUMarker.resetregion — Method

Reset the values stored using the region name regiontag. On success, true is returned.

LIKWID.GPUMarker.startregion — Method

Start measurements under the name regiontag. On success, true is returned.

LIKWID.GPUMarker.stopregion — Method

Stop measurements under the name regiontag. On success, true is returned.

LIKWID.GPUMarker.@gpuregion — Macro

Convenience macro for flanking code with GPUMarker.startregion and GPUMarker.stopregion.

Examples

julia> using LIKWID, CUDA

julia> GPUMarker.init()

julia> @gpuregion "sleeping..." sleep(1)
true

julia> @gpuregion "create rand vec" CUDA.rand(100)
true

julia> GPUMarker.close()