NVIDIA Monitoring (NvMon)

Note: This is a maturing feature. Only NVIDIA GPUs are supported.

Functions

LIKWID.NvMon.add_event_set — Method

add_event_set(estr) -> groupid

Add a performance group or a custom event set to the nvmon module. Returns a groupid which is required to later specify the event set.

LIKWID.NvMon.get_groups — Function

Return a list of all available nvmon groups for the GPU identified by gpuid.

Examples

julia> LIKWID.NvMon.get_groups()
4-element Vector{LIKWID.GroupInfoCompact}:
 DATA => Load to store ratio
 FLOPS_SP => Single-precision floating point
 FLOPS_HP => Half-precision floating point
 FLOPS_DP => Double-precision floating point

LIKWID.NvMon.get_id_of_active_group — Method

Return the groupid of the currently activate group.

LIKWID.NvMon.get_last_metric — Method

Return the derived metric result of the last measurement cycle identified by group groupid and the indices for metric metricidx and thread threadidx.

LIKWID.NvMon.get_last_result — Method

Return the raw counter register result of the last measurement cycle identified by group groupid and the indices for event eventidx and thread threadidx.

LIKWID.NvMon.get_longinfo_of_group — Method

Return the (long) description of a performance group.

LIKWID.NvMon.get_metric — Method

Return the derived metric result of all measurements identified by group groupid and the indices for metric metricidx and thread threadidx.

LIKWID.NvMon.get_name_of_counter — Method

Return the name of the counter register identified by groupid and eventidx.

LIKWID.NvMon.get_name_of_event — Method

Return the name of the event identified by groupid and eventidx.

LIKWID.NvMon.get_name_of_group — Method

Return the name of the group identified by groupid. If it is a custom event set, the name is set to Custom.

LIKWID.NvMon.get_name_of_metric — Method

Return the name of a derived metric identified by groupid and metricidx.

LIKWID.NvMon.get_number_of_events — Method

Return the number of events in the group.

LIKWID.NvMon.get_number_of_gpus — Method

Return the number of GPUs initialized in the nvmon module.

LIKWID.NvMon.get_number_of_groups — Method

Return the number of groups currently registered in the nvmon module.

LIKWID.NvMon.get_number_of_metrics — Method

Return the number of metrics in the group. Always zero for custom event sets.

LIKWID.NvMon.get_result — Method

Return the raw counter register result of all measurements identified by group groupid and the indices for event eventidx and thread threadidx.

LIKWID.NvMon.get_shortinfo_of_group — Method

Return the short information about a performance group.

LIKWID.NvMon.get_time_of_group — Method

Return the measurement time for group identified by groupid.

LIKWID.NvMon.read_counters — Method

Read the counter registers. To be executed after start_counters and before stop_counters. Returns true on success.

LIKWID.NvMon.setup_counters — Method

Program the counter registers to measure all events in group groupid. Returns true on success.

LIKWID.NvMon.start_counters — Method

Start the counter registers. Returns true on success.

LIKWID.NvMon.stop_counters — Method

Stop the counter registers. Returns true on success.

LIKWID.NvMon.switch_group — Method

Switch currently active group to groupid. Returns true on success.