Marker API (CPU)

Index

LIKWID.Marker.close

LIKWID.Marker.getregion

LIKWID.Marker.init

LIKWID.Marker.init_dynamic

LIKWID.Marker.init_nothreads

LIKWID.Marker.isactive

LIKWID.Marker.nextgroup

LIKWID.Marker.region

LIKWID.Marker.registerregion

LIKWID.Marker.resetregion

LIKWID.Marker.startregion

LIKWID.Marker.stopregion

LIKWID.Marker.threadinit

LIKWID.Marker.@parallelregion

LIKWID.Marker.@region

API

LIKWID.Marker.close — Method

Close the connection to the LIKWID Marker API and write out measurement data to file. This file will be evaluated by likwid-perfctr.

LIKWID.Marker.getregion — Method

getregion(regiontag::AbstractString, [num_events]) -> nevents, events, time, count

Get the intermediate results of the region identified by regiontag. On success, it returns * nevents: the number of events in the current group, * events: a list with all the aggregated event results, * time: the measurement time for the region and * count: the number of calls.

LIKWID.Marker.init — Method

Initialize the Marker API, assuming that julia is running under likwid-perfctr. Must be called previous to all other functions.

LIKWID.Marker.init_dynamic — Method

init_dynamic(group_or_groups; kwargs...)

Initialize the full Marker API from within the current Julia session (i.e. no likwird-perfctr necessary). A performance group, e.g. "FLOPS_DP", must be provided as the first argument.

LIKWID.Marker.init_nothreads — Method

Initialize the Marker API only on the main thread (assuming that julia is running under likwid-perfctr). LIKWID.Marker.threadinit() must be called manually.

LIKWID.Marker.isactive — Method

Checks whether the Marker API is active, i.e. julia has been started under likwid-perfctr -C ... -g ... -m.

LIKWID.Marker.nextgroup — Method

Switch to the next event set in a round-robin fashion. If you have set only one event set on the command line, this function performs no operation.

LIKWID.Marker.region — Method

region(f, regiontag::AbstractString)

Adds a LIKWID marker region around the execution of the given function f using Marker.startregion, Marker.stopregion under the hood. Note that LIKWID.Marker.init() and LIKWID.Marker.close() must be called before and after, respectively.

Examples

julia> using LIKWID

julia> Marker.init()

julia> region("sleeping...") do
           sleep(1)
       end
true

julia> region(()->rand(100), "create rand vec")
true

julia> Marker.close()

LIKWID.Marker.registerregion — Method

Register a region with name regiontag to the Marker API. On success, true is returned.

This is an optional function to reduce the overhead of region registration at Marker.startregion. If you don't call registerregion, the registration is done at startregion.

LIKWID.Marker.resetregion — Method

Reset the values stored using the region name regiontag. On success, true is returned.

LIKWID.Marker.startregion — Method

Start measurements under the name regiontag. On success, true is returned.

LIKWID.Marker.stopregion — Method

Stop measurements under the name regiontag. On success, true is returned.

LIKWID.Marker.threadinit — Method

Add the current thread to the Marker API.

LIKWID.Marker.@parallelregion — Macro

Convenience macro for flanking code with Marker.startregion and Marker.stopregion on all threads separately.

Examples

julia> using LIKWID

julia> Marker.init()

julia> @parallelregion begin
           Threads.@thread :static for i in 1:Threads.nthreads()
               # thread-local computation
           end
       end

julia> Marker.close()

LIKWID.Marker.@region — Macro

Convenience macro for flanking code with Marker.startregion and Marker.stopregion.

Examples

julia> using LIKWID

julia> Marker.init()

julia> @region "sleeping..." sleep(1)
true

julia> @region "create rand vec" rand(100)
true

julia> Marker.close()