Miscellaneous

Index

LIKWID.LIKWID_DEBUG

LIKWID.LIKWID_EVENTS

LIKWID.LIKWID_FILEPATH

LIKWID.LIKWID_FORCE

LIKWID.LIKWID_IGNORE_CPUSET

LIKWID.LIKWID_MODE

LIKWID.LIKWID_MPI_CONNECT

LIKWID.LIKWID_NO_ACCESS

LIKWID.LIKWID_PIN

LIKWID.LIKWID_SILENT

LIKWID.LIKWID_SKIP

LIKWID.LIKWID_THREADS

LIKWID.accessmode

LIKWID.clearenv

LIKWID.env

LIKWID.gpusupport

LIKWID.pinmask

LIKWID.setverbosity

API

LIKWID.setverbosity — Function

Set the verbosity level of the LIKWID library. Returns true on success.

Options are:

LIKWID.LibLikwid.DEBUGLEV_ONLY_ERROR or 0

LIKWID.LibLikwid.DEBUGLEV_INFO or 1

LIKWID.LibLikwid.DEBUGLEV_DETAIL or 2

LIKWID.LibLikwid.DEBUGLEV_DEVELOP or 3

LIKWID.gpusupport — Function

Returns whether LIKWID has been compiled with GPU support (i.e. has been compiled with NVIDIA_INTERFACE=true).

LIKWID.pinmask — Function

pinmask(N::Integer) -> mask

Generates a mask that can be supplied to likwid pin -s <mask> to pin N Julia threads.

Taken from https://discourse.julialang.org/t/thread-affinitization-pinning-julia-threads-to-cores/58069/8.

LIKWID.accessmode — Function

Query the access mode used by LIKWID, i.e. either ACCESSMODE_PERF, ACCESSMODE_DAEMON, or ACCESSMODE_DIRECT.

LIKWID.env — Function

List the values of LIKWID_* environment variables.

LIKWID.clearenv — Function

Unset all LIKWID_* environment variables (for the current session).

LIKWID.LIKWID_FORCE — Function

Enables the overwriting of counters that are detected to be in-use. The environment variable is similar to the -f/--force command line switch for likwid-perfctr.

LIKWID.LIKWID_NO_ACCESS — Function

The execution does not require the access layer (access to hardware counters). For example, this variable is set by likwid-topology or likwid-pin.

LIKWID.LIKWID_PIN — Function

The comma-separated list contains the CPUs the application threads should be pinned to. Careful, the first CPU in the cpuset must be the last entry because the application is pinned to this CPU per default.

LIKWID.LIKWID_SILENT — Function

Disable stdout output caused by the library and the scripts. Some scripts provide the -q/--quiet command line switch which provides the same functionality.

LIKWID.LIKWID_SKIP — Function

Variable content must be a hexmask. This hexmask describes which threads should be skipped while pinning. This function is required to avoid pinning the shepherd threads used by some OpenMP and MPI implementations. The version 4.3.1 introduced an automatic detection of the shepherd threads. In most cases the detection works, but if not, the hexmask overwrites the automatic detection.

LIKWID.LIKWID_DEBUG — Function

Verbosity settings for the LIKWID library.

LIKWID.LIKWID_IGNORE_CPUSET — Function

LIKWID respects the CPUset of the calling process. If you want to measure/run outside of this CPUset, use this environment variable. It will not ignore the CPUset but create a new CPUset internally which contains sysconf(_SC_NPROCESSORS_CONF) hardware threads.

LIKWID.LIKWID_FILEPATH — Function

Filepath for the result file of the MarkerAPI.

LIKWID.LIKWID_MODE — Function

Access mode for MarkerAPI. 1 is the code for the access daemon.

LIKWID.LIKWID_EVENTS — Function

Event string or performance group name. Multiple event strings or performance group names can be separated by |.

LIKWID.LIKWID_THREADS — Function

The CPUs LIKWID is configured to run on (comma-separated list).

LIKWID.LIKWID_MPI_CONNECT — Function

Connection method for Intel MPI. Default is ssh, see option -r of mpdboot or similar.