Preferences

By default, LIKWID.jl will assume that LIKWID is available as liblikwid on the system (i.e. present in LD_LOAD_PATH). You can overwrite this default and (permanently) specify the path to liblikwid as a Julia preference by using the tools below (LIKWID.Prefs.set_likwid in particular).

Index

LIKWID.Prefs.clear

LIKWID.Prefs.get_liblikwid

LIKWID.Prefs.set_liblikwid

Functions

LIKWID.Prefs.clear — Method

Clear all LIKWID.jl related preferences

LIKWID.Prefs.get_liblikwid — Method

Query the value of the liblikwid preference. Returns nothing if not set.

LIKWID.Prefs.set_liblikwid — Method

Set the path of the LIKWID library (liblikwid).