CPU Topology

The basis functionality of likwid-topology.

Example

julia> using LIKWID
julia> topo = LIKWID.get_cpu_topology()LIKWID.CpuTopology
├ numHWThreads: 2
├ activeHWThreads: 2
├ numSockets: 1
├ numDies: 0
├ numCoresPerSocket: 2
├ numThreadsPerCore: 1
├ numCacheLevels: 3
├ threadPool: ... (2 elements)
└ cacheLevels: ... (3 elements)
julia> topo.cacheLevels3-element Vector{LIKWID.CacheLevel}:
 LIKWID.CacheLevel(1, :data, 8, 64, 64, 32768, 1, 0)
 LIKWID.CacheLevel(2, :unified, 16, 1024, 64, 1048576, 1, 0)
 LIKWID.CacheLevel(3, :unified, 11, 53248, 64, 37486592, 2, 0)
julia> cpuinfo = LIKWID.get_cpu_info()LIKWID.CpuInfo
├ family: 6
├ model: 85
├ stepping: 4
├ vendor: 0
├ part: 0
├ clock: 0
├ turbo: false
├ osname: Intel(R) Xeon(R) Platinum 8171M CPU @ 2.60GHz
├ name: Intel Skylake SP processor
├ short_name: skylakeX
├ features: FP MMX SSE SSE2 HTT RDTSCP SSSE FMA SSE4.1 SSE4.2 AES AVX RDRAND HLE AVX2 RTM AVX512 RDSEED SSE3
├ isIntel: true
├ architecture: x86_64
├ supportUncore: true
├ supportClientmem: false
├ featureFlags: 4328454223
├ perf_version: 0
├ perf_num_ctr: 0
├ perf_width_ctr: 0
└ perf_num_fixed_ctr: 0

Graphical output

Currently, LIKWID.jl doesn't feature a native graphical visualization of the CPU topology. However, it provides a small "wrapper function" around likwid-topology -g which should give you an output like this:

julia> using LIKWID
julia> LIKWID.print_cpu_topology()Graphical Topology
********************************************************************************
Socket 0:
+---------------------+
| +-------+ +-------+ |
| |   0   | |   1   | |
| +-------+ +-------+ |
| +-------+ +-------+ |
| | 32 kB | | 32 kB | |
| +-------+ +-------+ |
| +-------+ +-------+ |
| |  1 MB | |  1 MB | |
| +-------+ +-------+ |
| +-----------------+ |
| |      36 MB      | |
| +-----------------+ |
+---------------------+

Functions

LIKWID.print_cpu_topology — Method

Graphical visualization of the CPU topology. Extracts the corresponding output of likwid-topology -g.

Types

LIKWID.CpuTopology — Type

CPU topology information

LIKWID.CpuInfo — Type

CPU information

LIKWID.HWThread — Type

Information about a hardware thread

LIKWID.CacheLevel — Type

Information about a cache level