Performance Monitoring (PerfMon)

The basis functionality of likwid-perfctr.

Example

(See https://github.com/JuliaPerf/LIKWID.jl/tree/main/examples/perfmon.)

# perfmon.jl
using LIKWID
using LinearAlgebra

A = rand(128, 64)
B = rand(64, 128)
C = zeros(128, 128)

# ncpus = LIKWID.get_cpu_topology().numCoresPerSocket
ncpus = 1
cpus = collect(0:ncpus-1)
LIKWID.PerfMon.init(cpus)
groupid = LIKWID.PerfMon.add_event_set("FLOPS_DP")
LIKWID.PerfMon.setup_counters(groupid)
LIKWID.PerfMon.start_counters()
for _ in 1:100
    mul!(C, A, B)
end
LIKWID.PerfMon.stop_counters()

LIKWID.PerfMon.start_counters()
for _ in 1:100
    mul!(C, A, B)
end
LIKWID.PerfMon.stop_counters()

for cpu in cpus
    @show cpu
    d = LIKWID.PerfMon.get_metric_results(groupid, cpu)
    display(d)
    println()
    d = LIKWID.PerfMon.get_event_results(groupid, cpu)
    display(d)
    println()
end
LIKWID.PerfMon.finalize()

Running julia perfmon.jl one obtains:

cpu = 0
OrderedCollections.OrderedDict{String, Float64} with 10 entries:
  "Runtime (RDTSC) [s]" => 0.0719716
  "Runtime unhalted [s]" => 0.0172482
  "Clock [MHz]" => 4585.47
  "CPI" => 1.83921
  "DP [MFLOP/s]" => 369.924
  "AVX DP [MFLOP/s]" => 0.0
  "AVX512 DP [MFLOP/s]" => 0.0
  "Packed [MUOPS/s]" => 184.962
  "Scalar [MUOPS/s]" => 0.0
  "Vectorization ratio" => 100.0
OrderedCollections.OrderedDict{String, Float64} with 7 entries:
  "INSTR_RETIRED_ANY" => 3.37623e7
  "CPU_CLK_UNHALTED_CORE" => 6.2096e7
  "CPU_CLK_UNHALTED_REF" => 4.87528e7
  "FP_ARITH_INST_RETIRED_128B_PACKED_DOUBLE" => 1.3312e7
  "FP_ARITH_INST_RETIRED_SCALAR_DOUBLE" => 0.0
  "FP_ARITH_INST_RETIRED_256B_PACKED_DOUBLE" => 0.0
  "FP_ARITH_INST_RETIRED_512B_PACKED_DOUBLE" => 0.0

All Functions

LIKWID.PerfMon.add_event_set — Method

add_event_set(estr) -> groupid

Add a performance group or a custom event set to the perfmon module. Returns a groupid which is required to later specify the event set.

LIKWID.PerfMon.get_event_results — Method

Get the name and results of all events of a given group for a given cpu thread.

LIKWID.PerfMon.get_groups — Method

Return a list of all available perfmon groups.

LIKWID.PerfMon.get_id_of_active_group — Method

Return the groupid of the currently activate group.

LIKWID.PerfMon.get_last_metric — Method

Return the derived metric result of the last measurement cycle identified by group groupid and the indices for metric metricidx and thread threadidx.

LIKWID.PerfMon.get_last_result — Method

Return the raw counter register result of the last measurement cycle identified by group groupid and the indices for event eventidx and thread threadidx.

LIKWID.PerfMon.get_longinfo_of_group — Method

Return the (long) description of a performance group.

LIKWID.PerfMon.get_metric — Method

Return the derived metric result of all measurements identified by group groupid and the indices for metric metricidx and thread threadidx.

LIKWID.PerfMon.get_metric_results — Method

Get the name and results of all metrics of a given group for a given cpu thread.

LIKWID.PerfMon.get_name_of_counter — Method

Return the name of the counter register identified by groupid and eventidx.

LIKWID.PerfMon.get_name_of_event — Method

Return the name of the event identified by groupid and eventidx.

LIKWID.PerfMon.get_name_of_group — Method

Return the name of the group identified by groupid. If it is a custom event set, the name is set to Custom.

LIKWID.PerfMon.get_name_of_metric — Method

Return the name of a derived metric identified by groupid and metricidx.

LIKWID.PerfMon.get_number_of_events — Method

Return the amount of events in the group.

LIKWID.PerfMon.get_number_of_groups — Method

Return the number of groups currently registered in the perfmon module.

LIKWID.PerfMon.get_number_of_metrics — Method

Return the amount of metrics in the group. Always zero for custom event sets.

LIKWID.PerfMon.get_number_of_threads — Method

Return the number of threads initialized in the perfmon module.

LIKWID.PerfMon.get_result — Method

Return the raw counter register result of all measurements identified by group groupid and the indices for event eventidx and thread threadidx.

LIKWID.PerfMon.get_shortinfo_of_group — Method

Return the short information about a performance group.

LIKWID.PerfMon.get_time_of_group — Method

Return the measurement time for group identified by groupid.

LIKWID.PerfMon.list_metrics — Method

List all the metrics of a given group.

LIKWID.PerfMon.read_counters — Method

Read the counter registers. To be executed after start_counters and before stop_counters. Returns true on success.

LIKWID.PerfMon.setup_counters — Method

Program the counter registers to measure all events in group groupid. Returns true on success.

LIKWID.PerfMon.start_counters — Method

Start the counter registers. Returns true on success.

LIKWID.PerfMon.stop_counters — Method

Stop the counter registers. Returns true on success.

LIKWID.PerfMon.switch_group — Method

Switch currently active group to groupid. Returns true on success.