Performance Monitoring (PerfMon)

The basis functionality of likwid-perfctr.

Example

(See https://github.com/JuliaPerf/LIKWID.jl/tree/main/examples/perfmon.)

# perfmon.jl
using LIKWID
using LinearAlgebra

A = rand(128, 64)
B = rand(64, 128)
C = zeros(128, 128)

cpu = 0 # starts with zero!
PerfMon.init(cpu)
groupid = PerfMon.add_event_set("FLOPS_DP")
PerfMon.setup_counters(groupid)

PerfMon.start_counters()
for _ in 1:100
    mul!(C, A, B)
end
PerfMon.stop_counters()

mdict = PerfMon.get_metric_results(groupid, cpu)
display(mdict)
println(); flush(stdout);
edict = PerfMon.get_event_results(groupid, cpu)
display(edict)

PerfMon.finalize()

Running the above with julia perfmon.jl one obtains (modulo architectural differences):

OrderedCollections.OrderedDict{String, Float64} with 10 entries:
  "Runtime (RDTSC) [s]" => 0.695678
  "Runtime unhalted [s]" => 0.0014145
  "Clock [MHz]" => 1200.12
  "CPI" => 1.00722
  "DP [MFLOP/s]" => 38.2706
  "AVX DP [MFLOP/s]" => 38.2706
  "AVX512 DP [MFLOP/s]" => 38.2706
  "Packed [MUOPS/s]" => 4.78382
  "Scalar [MUOPS/s]" => 0.0
  "Vectorization ratio" => 100.0
OrderedCollections.OrderedDict{String, Float64} with 7 entries:
  "INSTR_RETIRED_ANY" => 4.63437e6
  "CPU_CLK_UNHALTED_CORE" => 4.66782e6
  "CPU_CLK_UNHALTED_REF" => 1.28352e7
  "FP_ARITH_INST_RETIRED_128B_PACKED_DOUBLE" => 0.0
  "FP_ARITH_INST_RETIRED_SCALAR_DOUBLE" => 0.0
  "FP_ARITH_INST_RETIRED_256B_PACKED_DOUBLE" => 0.0
  "FP_ARITH_INST_RETIRED_512B_PACKED_DOUBLE" => 3.328e6

API

LIKWID.PerfMon.add_event_set

LIKWID.PerfMon.get_event_results

LIKWID.PerfMon.get_groups

LIKWID.PerfMon.get_id_of_active_group

LIKWID.PerfMon.get_id_of_event

LIKWID.PerfMon.get_id_of_group

LIKWID.PerfMon.get_id_of_metric

LIKWID.PerfMon.get_last_metric

LIKWID.PerfMon.get_last_result

LIKWID.PerfMon.get_longinfo_of_group

LIKWID.PerfMon.get_metric

LIKWID.PerfMon.get_metric_results

LIKWID.PerfMon.get_metric_results

LIKWID.PerfMon.get_name_of_counter

LIKWID.PerfMon.get_name_of_event

LIKWID.PerfMon.get_name_of_group

LIKWID.PerfMon.get_name_of_metric

LIKWID.PerfMon.get_number_of_events

LIKWID.PerfMon.get_number_of_groups

LIKWID.PerfMon.get_number_of_metrics

LIKWID.PerfMon.get_number_of_threads

LIKWID.PerfMon.get_result

LIKWID.PerfMon.get_shortinfo_of_group

LIKWID.PerfMon.get_time_of_group

LIKWID.PerfMon.list_events

LIKWID.PerfMon.list_metrics

LIKWID.PerfMon.perfmon

LIKWID.PerfMon.read_counters

LIKWID.PerfMon.setup_counters

LIKWID.PerfMon.start_counters

LIKWID.PerfMon.stop_counters

LIKWID.PerfMon.switch_group

LIKWID.GroupInfoCompact

Functions

LIKWID.PerfMon.add_event_set — Method

add_event_set(estr) -> groupid

Add a performance group or a custom event set to the perfmon module. Returns a groupid (starting at 1) which is required to later specify the event set.

LIKWID.PerfMon.get_event_results — Method

get_event_results([groupid_or_groupname, eventid_or_eventname, threadid::Integer])

Retrieve the results of monitored events. Same as get_metric_results but for raw events.

LIKWID.PerfMon.get_groups — Method

Return a list of all available perfmon groups.

Examples

julia> PerfMon.get_groups()
30-element Vector{LIKWID.GroupInfoCompact}:
 MEM_SP => Overview of arithmetic and main memory performance
 CYCLE_ACTIVITY => Cycle Activities
 ENERGY => Power and Energy consumption
 UOPS_RETIRE => UOPs retirement
 PMM => Intel Optane DC bandwidth in MBytes/s
 TLB_INSTR => L1 Instruction TLB miss rate/ratio
 DATA => Load to store ratio
 UOPS_ISSUE => UOPs issueing
 L2CACHE => L2 cache miss rate/ratio
 L2 => L2 cache bandwidth in MBytes/s
 ⋮
 TMA => Top down cycle allocation
 CLOCK => Power and Energy consumption
 FLOPS_DP => Double Precision MFLOP/s
 CYCLE_STALLS => Cycle Activities (Stalls)
 L3 => L3 cache bandwidth in MBytes/s
 UPI => UPI traffic
 L3NEW => L3 cache bandwidth in MBytes/s
 L3PF => L3 cache bandwidth in MBytes/s
 L2L3 => L3 cache bandwidth in MBytes/s

LIKWID.PerfMon.get_id_of_active_group — Method

Return the groupid of the currently activate group.

LIKWID.PerfMon.get_id_of_event — Method

Get the id of the event with the given name.

LIKWID.PerfMon.get_id_of_group — Method

Get the id of the group with the given name.

LIKWID.PerfMon.get_id_of_metric — Method

Get the id of the metric with the given name.

LIKWID.PerfMon.get_last_metric — Method

Return the derived metric result of the last measurement cycle identified by group groupid and the indices for metric metricidx and thread threadidx (all starting at 1).

LIKWID.PerfMon.get_last_result — Method

Return the raw counter register result of the last measurement cycle identified by group groupid and the indices for event eventidx and thread threadidx (all starting at 1).

LIKWID.PerfMon.get_longinfo_of_group — Method

Return the (long) description of a performance group with id groupid (starts at 1).

LIKWID.PerfMon.get_metric — Method

Return the derived metric result of all measurements identified by group groupid and the indices for metric metricidx and thread threadidx (all starting at 1).

LIKWID.PerfMon.get_metric_results — Method

get_metric_results([groupid_or_groupname, metricid_or_metricname, threadid::Integer])

Retrieve the results of monitored metrics.

Optionally, a group, metric, and threadid can be provided to select a subset of metrics or a single metric. If given as integers, note that groupid, metricid, and threadid all start at 1 and the latter enumerates the monitored cpu threads.

If no arguments are provided, a nested data structure is returned in which different levels correspond to performance groups, cpu threads, and metrics (in this order).

Examples

julia> PerfMon.get_metric_results("FLOPS_DP")
4-element Vector{OrderedDict{String, Float64}}:
 OrderedDict("Runtime (RDTSC) [s]" => 1.1381168037989857, "Runtime unhalted [s]" => 0.0016642799007831007, "Clock [MHz]" => 2911.9285695819794, "CPI" => NaN, "DP [MFLOP/s]" => 0.0)
 OrderedDict("Runtime (RDTSC) [s]" => 1.1381168037989857, "Runtime unhalted [s]" => 1.4755564705029072, "Clock [MHz]" => 3523.1114993407705, "CPI" => 0.3950777002592585, "DP [MFLOP/s]" => 17608.069202657578)
 OrderedDict("Runtime (RDTSC) [s]" => 1.1381168037989857, "Runtime unhalted [s]" => 7.80437228993214e-5, "Clock [MHz]" => 2638.6244625814124, "CPI" => NaN, "DP [MFLOP/s]" => 0.0)
 OrderedDict("Runtime (RDTSC) [s]" => 1.1381168037989857, "Runtime unhalted [s]" => 7.050705084934875e-5, "Clock [MHz]" => 2807.7525945849698, "CPI" => NaN, "DP [MFLOP/s]" => 0.0)

julia> PerfMon.get_metric_results("FLOPS_DP", 2) # results of second monitored cpu thread
OrderedDict{String, Float64} with 5 entries:
  "Runtime (RDTSC) [s]"  => 1.13812
  "Runtime unhalted [s]" => 1.47556
  "Clock [MHz]"          => 3523.11
  "CPI"                  => 0.395078
  "DP [MFLOP/s]"         => 17608.1

julia> PerfMon.get_metric_results("FLOPS_DP", "DP [MFLOP/s]", 2)
17608.069202657578

LIKWID.PerfMon.get_metric_results — Method

get_metric_results()

Get the metric results for all performance groups and all monitored (PerfMon.init) cpu threads.

Returns a an OrderedDict whose keys correspond to the performance groups and the values hold the results for all monitored cpu threads.

Examples

julia> results = PerfMon.get_metric_results()
OrderedDict{String, Vector{OrderedDict{String, Float64}}} with 1 entry:
  "FLOPS_DP" => [OrderedDict("Runtime (RDTSC) [s]"=>1.13812, "Runtime unhalted [s]"=>0.00166428, "Clock [MHz]"=>291…

julia> PerfMon.get_metric_results()["FLOPS_DP"][2]["DP [MFLOP/s]"]
17608.069202657578

LIKWID.PerfMon.get_name_of_counter — Method

Return the name of the counter register identified by groupid and eventidx (both starting at 1).

LIKWID.PerfMon.get_name_of_event — Method

Return the name of the event identified by groupid and eventidx (both starting at 1).

LIKWID.PerfMon.get_name_of_group — Method

Return the name of the group identified by groupid (starts at 1). If it is a custom event set, the name is set to Custom.

LIKWID.PerfMon.get_name_of_metric — Method

Return the name of a derived metric identified by groupid and metricidx (both starting at 1).

LIKWID.PerfMon.get_number_of_events — Method

Return the amount of events in the given group with id groupid (starts at 1).

LIKWID.PerfMon.get_number_of_groups — Method

Return the number of groups currently registered in the perfmon module.

LIKWID.PerfMon.get_number_of_metrics — Method

Return the amount of metrics in the given group with id groupid (starts at 1). Always zero for custom event sets.

LIKWID.PerfMon.get_number_of_threads — Method

Return the number of threads initialized in the perfmon module.

LIKWID.PerfMon.get_result — Method

Return the raw counter register result of all measurements identified by group groupid and the indices for event eventidx and thread threadidx (all starting at 1).

LIKWID.PerfMon.get_shortinfo_of_group — Method

Return the short information about a performance group with id groupid (starts at 1).

LIKWID.PerfMon.get_time_of_group — Method

Return the measurement time for group identified by groupid (starts at 1).

LIKWID.PerfMon.list_events — Method

List all the events of a given group (groupid starts at 1).

LIKWID.PerfMon.list_metrics — Method

List all the metrics of a given group (groupid starts at 1).

LIKWID.PerfMon.perfmon — Method

perfmon(f, group_or_groups[; cpuids, autopin=true])

Monitor performance groups while executing the given function f on one or multiple Julia threads. Note that

PerfMon.init and PerfMon.finalize() are called automatically, and

the measurement of multiple performance groups is sequential and f is executed multiple times.

Keyword arguments:

cpuids (default: currently used CPU threads): specify the CPU threads (~ cores) to be monitored

autopin (default: true): automatically pin Julia threads to the CPU threads (~ cores) they are currently running on (to avoid migration and wrong results).

Example

julia> using LIKWID

julia> x = rand(1000); y = rand(1000);

julia> metrics, events = perfmon("FLOPS_DP") do
           x .+ y;
       end;

julia> metrics                                                                      
OrderedDict{String, Float64} with 5 entries:
  "Runtime (RDTSC) [s]"  => 8.56091e-6
  "Runtime unhalted [s]" => 3.22377e-5
  "Clock [MHz]"          => 3506.47
  "CPI"                  => 4.78484
  "DP [MFLOP/s]"         => 116.81
                                          
julia> events                                                                       
OrderedDict{String, Float64} with 6 entries:
  "ACTUAL_CPU_CLOCK"          => 78974.0
  "MAX_CPU_CLOCK"             => 55174.0
  "RETIRED_INSTRUCTIONS"      => 5977.0
  "CPU_CLOCKS_UNHALTED"       => 28599.0
  "RETIRED_SSE_AVX_FLOPS_ALL" => 1000.0
  "MERGE"                     => 0.0

julia> metrics, events = perfmon(("FLOPS_DP", "MEM1")) do
           x .+ y;
       end;

LIKWID.PerfMon.read_counters — Method

Read the counter registers. To be executed after start_counters and before stop_counters. Returns true on success.

LIKWID.PerfMon.setup_counters — Method

Program the counter registers to measure all events in group groupid (starts at 1). Returns true on success.

LIKWID.PerfMon.start_counters — Method

Start the counter registers. Returns true on success.

LIKWID.PerfMon.stop_counters — Method

Stop the counter registers. Returns true on success.

LIKWID.PerfMon.switch_group — Method

Switch currently active group to groupid (starts with 1). Returns true on success.

LIKWID.PerfMon.@perfmon — Macro

@perfmon groupname codeblock

See also: perfmon

Example

julia> using LIKWID

julia> x = rand(1000); y = rand(1000);

julia> metrics, events = @perfmon "FLOPS_DP" x .+ y;

julia> metrics                                                                      
OrderedDict{String, Float64} with 5 entries:
  "Runtime (RDTSC) [s]"  => 8.56091e-6
  "Runtime unhalted [s]" => 3.22377e-5
  "Clock [MHz]"          => 3506.47
  "CPI"                  => 4.78484
  "DP [MFLOP/s]"         => 116.81
                                          
julia> events                                                                       
OrderedDict{String, Float64} with 6 entries:
  "ACTUAL_CPU_CLOCK"          => 78974.0
  "MAX_CPU_CLOCK"             => 55174.0
  "RETIRED_INSTRUCTIONS"      => 5977.0
  "CPU_CLOCKS_UNHALTED"       => 28599.0
  "RETIRED_SSE_AVX_FLOPS_ALL" => 1000.0
  "MERGE"                     => 0.0

Types

LIKWID.GroupInfoCompact — Type

Essential information about a performance group