Performance Monitoring (PerfMon)

The basis functionality of likwid-perfctr.

Example

(See https://github.com/JuliaPerf/LIKWID.jl/tree/main/examples/perfmon.)

# perfmon.jl
using LIKWID
using LinearAlgebra

A = rand(128, 64)
B = rand(64, 128)
C = zeros(128, 128)

cpu = 0 # starts with zero!
LIKWID.PerfMon.init(cpu)
groupid = LIKWID.PerfMon.add_event_set("FLOPS_DP")
LIKWID.PerfMon.setup_counters(groupid)

LIKWID.PerfMon.start_counters()
for _ in 1:100
    mul!(C, A, B)
end
LIKWID.PerfMon.stop_counters()

mdict = LIKWID.PerfMon.get_metric_results(groupid, cpu)
display(mdict)
println(); flush(stdout);
edict = LIKWID.PerfMon.get_event_results(groupid, cpu)
display(edict)

LIKWID.PerfMon.finalize()

Running the above with julia perfmon.jl one obtains (modulo architectural differences):

OrderedCollections.OrderedDict{String, Float64} with 10 entries:
  "Runtime (RDTSC) [s]" => 0.447041
  "Runtime unhalted [s]" => 0.00339698
  "Clock [MHz]" => 4597.86
  "CPI" => 0.494021
  "DP [MFLOP/s]" => 59.5561
  "AVX DP [MFLOP/s]" => 0.0
  "AVX512 DP [MFLOP/s]" => 0.0
  "Packed [MUOPS/s]" => 29.778
  "Scalar [MUOPS/s]" => 7.82926e-5
  "Vectorization ratio" => 99.9997
OrderedCollections.OrderedDict{String, Float64} with 7 entries:
  "INSTR_RETIRED_ANY" => 2.47543e7
  "CPU_CLK_UNHALTED_CORE" => 1.22292e7
  "CPU_CLK_UNHALTED_REF" => 9.5751e6
  "FP_ARITH_INST_RETIRED_128B_PACKED_DOUBLE" => 1.3312e7
  "FP_ARITH_INST_RETIRED_SCALAR_DOUBLE" => 35.0
  "FP_ARITH_INST_RETIRED_256B_PACKED_DOUBLE" => 0.0
  "FP_ARITH_INST_RETIRED_512B_PACKED_DOUBLE" => 0.0

Functions

LIKWID.PerfMon.add_event_set — Method

add_event_set(estr) -> groupid

Add a performance group or a custom event set to the perfmon module. Returns a groupid which is required to later specify the event set.

LIKWID.PerfMon.get_event_results — Method

Get the name and results of all events of a given group for a given cpu thread.

LIKWID.PerfMon.get_groups — Method

Return a list of all available perfmon groups.

Examples

julia> LIKWID.PerfMon.get_groups()
30-element Vector{LIKWID.GroupInfoCompact}:
 MEM_SP => Overview of arithmetic and main memory performance
 CYCLE_ACTIVITY => Cycle Activities
 ENERGY => Power and Energy consumption
 UOPS_RETIRE => UOPs retirement
 PMM => Intel Optane DC bandwidth in MBytes/s
 TLB_INSTR => L1 Instruction TLB miss rate/ratio
 DATA => Load to store ratio
 UOPS_ISSUE => UOPs issueing
 L2CACHE => L2 cache miss rate/ratio
 L2 => L2 cache bandwidth in MBytes/s
 ⋮
 TMA => Top down cycle allocation
 CLOCK => Power and Energy consumption
 FLOPS_DP => Double Precision MFLOP/s
 CYCLE_STALLS => Cycle Activities (Stalls)
 L3 => L3 cache bandwidth in MBytes/s
 UPI => UPI traffic
 L3NEW => L3 cache bandwidth in MBytes/s
 L3PF => L3 cache bandwidth in MBytes/s
 L2L3 => L3 cache bandwidth in MBytes/s

LIKWID.PerfMon.get_id_of_active_group — Method

Return the groupid of the currently activate group.

LIKWID.PerfMon.get_last_metric — Method

Return the derived metric result of the last measurement cycle identified by group groupid and the indices for metric metricidx and thread threadidx.

LIKWID.PerfMon.get_last_result — Method

Return the raw counter register result of the last measurement cycle identified by group groupid and the indices for event eventidx and thread threadidx.

LIKWID.PerfMon.get_longinfo_of_group — Method

Return the (long) description of a performance group.

LIKWID.PerfMon.get_metric — Method

Return the derived metric result of all measurements identified by group groupid and the indices for metric metricidx and thread threadidx.

LIKWID.PerfMon.get_metric_results — Method

Get the name and results of all metrics of a given group for a given cpu thread.

LIKWID.PerfMon.get_name_of_counter — Method

Return the name of the counter register identified by groupid and eventidx.

LIKWID.PerfMon.get_name_of_event — Method

Return the name of the event identified by groupid and eventidx.

LIKWID.PerfMon.get_name_of_group — Method

Return the name of the group identified by groupid. If it is a custom event set, the name is set to Custom.

LIKWID.PerfMon.get_name_of_metric — Method

Return the name of a derived metric identified by groupid and metricidx.

LIKWID.PerfMon.get_number_of_events — Method

Return the amount of events in the group.

LIKWID.PerfMon.get_number_of_groups — Method

Return the number of groups currently registered in the perfmon module.

LIKWID.PerfMon.get_number_of_metrics — Method

Return the amount of metrics in the group. Always zero for custom event sets.

LIKWID.PerfMon.get_number_of_threads — Method

Return the number of threads initialized in the perfmon module.

LIKWID.PerfMon.get_result — Method

Return the raw counter register result of all measurements identified by group groupid and the indices for event eventidx and thread threadidx.

LIKWID.PerfMon.get_shortinfo_of_group — Method

Return the short information about a performance group.

LIKWID.PerfMon.get_time_of_group — Method

Return the measurement time for group identified by groupid.

LIKWID.PerfMon.list_metrics — Method

List all the metrics of a given group.

LIKWID.PerfMon.read_counters — Method

Read the counter registers. To be executed after start_counters and before stop_counters. Returns true on success.

LIKWID.PerfMon.setup_counters — Method

Program the counter registers to measure all events in group groupid. Returns true on success.

LIKWID.PerfMon.start_counters — Method

Start the counter registers. Returns true on success.

LIKWID.PerfMon.stop_counters — Method

Stop the counter registers. Returns true on success.

LIKWID.PerfMon.switch_group — Method

Switch currently active group to groupid. Returns true on success.