CPU / NUMA Topology

The basis functionality of likwid-topology.

Example

Query CPU topology information:

julia> topo = LIKWID.get_cpu_topology()LIKWID.CpuTopology
├ numHWThreads: 48
├ activeHWThreads: 48
├ numSockets: 2
├ numDies: 0
├ numCoresPerSocket: 12
├ numThreadsPerCore: 2
├ numCacheLevels: 3
├ threadPool: ... (48 elements)
└ cacheLevels: ... (3 elements)
julia> topo.threadPool48-element Vector{LIKWID.HWThread}:
 LIKWID.HWThread(0, 0, 0, 0, 0, 1)
 LIKWID.HWThread(0, 1, 0, 1, 0, 1)
 LIKWID.HWThread(0, 2, 0, 2, 0, 1)
 LIKWID.HWThread(0, 3, 0, 3, 0, 1)
 LIKWID.HWThread(0, 4, 0, 4, 0, 1)
 LIKWID.HWThread(0, 5, 0, 5, 0, 1)
 LIKWID.HWThread(0, 6, 0, 6, 0, 1)
 LIKWID.HWThread(0, 7, 0, 7, 0, 1)
 LIKWID.HWThread(0, 8, 0, 8, 0, 1)
 LIKWID.HWThread(0, 9, 0, 9, 0, 1)
 ⋮
 LIKWID.HWThread(1, 15, 1, 39, 0, 1)
 LIKWID.HWThread(1, 16, 1, 40, 0, 1)
 LIKWID.HWThread(1, 17, 1, 41, 0, 1)
 LIKWID.HWThread(1, 18, 1, 42, 0, 1)
 LIKWID.HWThread(1, 19, 1, 43, 0, 1)
 LIKWID.HWThread(1, 20, 1, 44, 0, 1)
 LIKWID.HWThread(1, 21, 1, 45, 0, 1)
 LIKWID.HWThread(1, 22, 1, 46, 0, 1)
 LIKWID.HWThread(1, 23, 1, 47, 0, 1)
julia> topo.cacheLevels3-element Vector{LIKWID.CacheLevel}:
 LIKWID.CacheLevel(1, :data, 8, 64, 64, 32768, 2, 0)
 LIKWID.CacheLevel(2, :unified, 16, 1024, 64, 1048576, 2, 0)
 LIKWID.CacheLevel(3, :unified, 11, 36864, 64, 25952256, 24, 0)

Get detailed CPU information:

julia> cpuinfo = LIKWID.get_cpu_info()LIKWID.CpuInfo
├ family: 6
├ model: 85
├ stepping: 7
├ vendor: 0
├ part: 0
├ clock: 0
├ turbo: true
├ osname: Intel(R) Xeon(R) Gold 6246 CPU @ 3.30GHz
├ name: Intel Cascadelake SP processor
├ short_name: CLX
├ features: FP ACPI MMX SSE SSE2 HTT TM RDTSCP MONITOR VMX EIST TM2 SSSE FMA SSE4.1 SSE4.2 AES AVX RDRAND AVX2 AVX512 RDSEED SSE3
├ isIntel: true
├ architecture: x86_64
├ supportUncore: true
├ supportClientmem: false
├ featureFlags: 4325310463
├ perf_version: 4
├ perf_num_ctr: 4
├ perf_width_ctr: 48
└ perf_num_fixed_ctr: 3

Query information about NUMA nodes:

julia> numa = LIKWID.get_numa_topology()LIKWID.NumaTopology
├ numberOfNodes: 2
└ nodes: ... (2 elements)
julia> numa.nodes2-element Vector{LIKWID.NumaNode}:
 NumaNode()
 NumaNode()
julia> numa_node = first(numa.nodes)LIKWID.NumaNode
├ id: 0
├ totalMemory: 62.59 GB
├ freeMemory: 62.07 GB
├ numberOfProcessors: 24
├ processors: [0, 24, 1, 25, 2, 26, 3, 27, 4, 28  …  7, 31, 8, 32, 9, 33, 10, 34, 11, 35]
├ numberOfDistances: 2
└ distances: [10, 21]

Graphical output

Currently, LIKWID.jl doesn't feature a native graphical visualization of the CPU topology. However, it provides a small "wrapper function" around likwid-topology -g which should give you an output like this:

julia> LIKWID.print_cpu_topology()┌ Warning: CUDA cannot be found and initialized (cuInit failed).
│ ERROR - [./src/topology_gpu.c:topology_gpu_init:232] Cannot get number of devices from CUDA library
└ @ LIKWID /home/saturn/unrz/unrz139/gitlab-runner/builds/0/ub55yzis/LIKWID.jl/src/topology.jl:161
Graphical Topology
********************************************************************************
Socket 0:
+-------------------------------------------------------------------------------------------------------------------------------------+
| +--------+ +--------+ +--------+ +--------+ +--------+ +--------+ +--------+ +--------+ +--------+ +--------+ +--------+ +--------+ |
| |  0 24  | |  1 25  | |  2 26  | |  3 27  | |  4 28  | |  5 29  | |  6 30  | |  7 31  | |  8 32  | |  9 33  | | 10 34  | | 11 35  | |
| +--------+ +--------+ +--------+ +--------+ +--------+ +--------+ +--------+ +--------+ +--------+ +--------+ +--------+ +--------+ |
| +--------+ +--------+ +--------+ +--------+ +--------+ +--------+ +--------+ +--------+ +--------+ +--------+ +--------+ +--------+ |
| |  32 kB | |  32 kB | |  32 kB | |  32 kB | |  32 kB | |  32 kB | |  32 kB | |  32 kB | |  32 kB | |  32 kB | |  32 kB | |  32 kB | |
| +--------+ +--------+ +--------+ +--------+ +--------+ +--------+ +--------+ +--------+ +--------+ +--------+ +--------+ +--------+ |
| +--------+ +--------+ +--------+ +--------+ +--------+ +--------+ +--------+ +--------+ +--------+ +--------+ +--------+ +--------+ |
| |  1 MB  | |  1 MB  | |  1 MB  | |  1 MB  | |  1 MB  | |  1 MB  | |  1 MB  | |  1 MB  | |  1 MB  | |  1 MB  | |  1 MB  | |  1 MB  | |
| +--------+ +--------+ +--------+ +--------+ +--------+ +--------+ +--------+ +--------+ +--------+ +--------+ +--------+ +--------+ |
| +---------------------------------------------------------------------------------------------------------------------------------+ |
| |                                                              25 MB                                                              | |
| +---------------------------------------------------------------------------------------------------------------------------------+ |
+-------------------------------------------------------------------------------------------------------------------------------------+
Socket 1:
+-------------------------------------------------------------------------------------------------------------------------------------+
| +--------+ +--------+ +--------+ +--------+ +--------+ +--------+ +--------+ +--------+ +--------+ +--------+ +--------+ +--------+ |
| | 12 36  | | 13 37  | | 14 38  | | 15 39  | | 16 40  | | 17 41  | | 18 42  | | 19 43  | | 20 44  | | 21 45  | | 22 46  | | 23 47  | |
| +--------+ +--------+ +--------+ +--------+ +--------+ +--------+ +--------+ +--------+ +--------+ +--------+ +--------+ +--------+ |
| +--------+ +--------+ +--------+ +--------+ +--------+ +--------+ +--------+ +--------+ +--------+ +--------+ +--------+ +--------+ |
| |  32 kB | |  32 kB | |  32 kB | |  32 kB | |  32 kB | |  32 kB | |  32 kB | |  32 kB | |  32 kB | |  32 kB | |  32 kB | |  32 kB | |
| +--------+ +--------+ +--------+ +--------+ +--------+ +--------+ +--------+ +--------+ +--------+ +--------+ +--------+ +--------+ |
| +--------+ +--------+ +--------+ +--------+ +--------+ +--------+ +--------+ +--------+ +--------+ +--------+ +--------+ +--------+ |
| |  1 MB  | |  1 MB  | |  1 MB  | |  1 MB  | |  1 MB  | |  1 MB  | |  1 MB  | |  1 MB  | |  1 MB  | |  1 MB  | |  1 MB  | |  1 MB  | |
| +--------+ +--------+ +--------+ +--------+ +--------+ +--------+ +--------+ +--------+ +--------+ +--------+ +--------+ +--------+ |
| +---------------------------------------------------------------------------------------------------------------------------------+ |
| |                                                              25 MB                                                              | |
| +---------------------------------------------------------------------------------------------------------------------------------+ |
+-------------------------------------------------------------------------------------------------------------------------------------+

API

LIKWID.finalize_numa

LIKWID.finalize_topology

LIKWID.get_cpu_info

LIKWID.get_cpu_topology

LIKWID.get_numa_topology

LIKWID.init_numa

LIKWID.init_topology

LIKWID.print_supported_cpus

LIKWID.CacheLevel

LIKWID.CpuInfo

LIKWID.CpuTopology

LIKWID.HWThread

LIKWID.NumaNode

LIKWID.NumaTopology

Functions

LIKWID.init_topology — Function

Initialize LIKWIDs topology module.

LIKWID.finalize_topology — Function

Close and finalize LIKWIDs topology module.

LIKWID.get_cpu_topology — Function

get_cpu_topology() -> CpuTopology

Get the CPU topology of the machine.

Automatically initializes the topology and NUMA modules, i.e. calls LIKWID.init_topology and LIKWID.init_numa.

LIKWID.get_cpu_info — Function

get_cpu_info() -> CpuInfo

Get detailed information about the CPU.

Automatically initializes the topology and NUMA modules, i.e. calls LIKWID.init_topology and LIKWID.init_numa.

LIKWID.print_supported_cpus — Function

print_supported_cpus(; cprint=true)

Print a list of all supported CPUs.

If cprint=false, LIKWID.jl will first capture the stdout and then print the list.

LIKWID.init_numa — Function

Initialize LIKWIDs NUMA module.

LIKWID.finalize_numa — Function

Close and finalize LIKWIDs NUMA module.

LIKWID.get_numa_topology — Function

get_numa_topology() -> NumaTopology

Get the NUMA topology of the machine.

Automatically initializes the topology, NUMA, and affinity modules, i.e. calls LIKWID.init_topology, LIKWID.init_numa, and LIKWID.init_affinity.

Types

LIKWID.CpuTopology — Type

CPU topology information

LIKWID.CpuInfo — Type

CPU information

LIKWID.HWThread — Type

Information about a hardware thread

LIKWID.CacheLevel — Type

Information about a cache level

LIKWID.NumaTopology — Type

CPU topology information

LIKWID.NumaNode — Type

Information about a NUMA node