CPU / NUMA Topology

The basis functionality of likwid-topology.

Example

Query CPU topology information:

julia> topo = LIKWID.get_cpu_topology()ERROR: UndefVarError: LIKWID not defined
julia> topo.threadPoolERROR: UndefVarError: topo not defined
julia> topo.cacheLevelsERROR: UndefVarError: topo not defined

Get detailed CPU information:

julia> cpuinfo = LIKWID.get_cpu_info()ERROR: UndefVarError: LIKWID not defined

Query information about NUMA nodes:

julia> numa = LIKWID.get_numa_topology()ERROR: UndefVarError: LIKWID not defined
julia> numa.nodesERROR: UndefVarError: numa not defined
julia> numa_node = first(numa.nodes)ERROR: UndefVarError: numa not defined

Graphical output

Currently, LIKWID.jl doesn't feature a native graphical visualization of the CPU topology. However, it provides a small "wrapper function" around likwid-topology -g which should give you an output like this:

julia> LIKWID.print_cpu_topology()ERROR: UndefVarError: LIKWID not defined

Functions

LIKWID.init_topology — Method

Initialize LIKWIDs topology module.

LIKWID.finalize_topology — Method

Close and finalize LIKWIDs topology module.

LIKWID.get_cpu_topology — Method

get_cpu_topology() -> CpuTopology

Get the CPU topology of the machine.

Automatically initializes the topology and NUMA modules, i.e. calls init_topology and init_numa.

LIKWID.get_cpu_info — Method

get_cpu_info() -> CpuInfo

Get detailed information about the CPU.

Automatically initializes the topology and NUMA modules, i.e. calls init_topology and init_numa.

LIKWID.print_supported_cpus — Method

print_supported_cpus(; cprint=true)

Print a list of all supported CPUs.

If cprint=false, LIKWID.jl will first capture the stdout and then print the list.

LIKWID.init_numa — Method

Initialize LIKWIDs NUMA module.

LIKWID.finalize_numa — Method

Close and finalize LIKWIDs NUMA module.

LIKWID.get_numa_topology — Method

get_numa_topology() -> NumaTopology

Get the NUMA topology of the machine.

Automatically initializes the topology, NUMA, and affinity modules, i.e. calls init_topology, init_numa, and init_affinity.

Types

LIKWID.CpuTopology — Type

CPU topology information

LIKWID.CpuInfo — Type

CPU information

LIKWID.HWThread — Type

Information about a hardware thread

LIKWID.CacheLevel — Type

Information about a cache level

LIKWID.NumaTopology — Type

CPU topology information

LIKWID.NumaNode — Type

Information about a NUMA node

Supported CPUs

julia> LIKWID.print_supported_cpus()ERROR: UndefVarError: LIKWID not defined